moved ui utils to root of ayon_common
